Tiny improvement
