+++ tests/Integration/CacheTest.php
@@ -5,12 +5,15 @@ namespace uuf6429\PhpCsFixerBlockstringTests\Integration;
 use JsonException;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Process\Process;
+use uuf6429\PhpCsFixerBlockstringTests\CustomAssertionsTrait;
 
 /**
  * @internal
  */
 final class CacheTest extends TestCase
 {
+	use CustomAssertionsTrait;
+
 	private const PCF_BINARY_PATH = __DIR__ . '/../../vendor/bin/php-cs-fixer';
 
 	private static string $workspace;
@@ -79,45 +82,5 @@ final class CacheTest extends TestCase
 			self::$cacheFile,
 			['TEST_FORMATTER_CACHE_FINGERPRINT' => $cacheFingerprint]
 		);
-	}
-
-	/**
-	 * @param array<string, string> $placeholders
-	 * @throws JsonException
-	 */
-	private function assertJsonFileWithinJsonFile(string $expectedFile, string $actualSubsetFile, array $placeholders = []): void
-	{
-		$this->assertFileExists($expectedFile);
-		$this->assertFileExists($actualSubsetFile);
-		$this->assertArraySubsetRecursive(
-			(array)json_decode(strtr((string)file_get_contents($expectedFile), $placeholders), true, 512, JSON_THROW_ON_ERROR),
-			(array)json_decode((string)file_get_contents($actualSubsetFile), true, 512, JSON_THROW_ON_ERROR),
-			sprintf(
-				'Failed asserting that %s contains %s',
-				(string)realpath($actualSubsetFile),
-				(string)realpath($expectedFile)
-			)
-		);
-	}
-
-	/**
-	 * @param array<array-key, mixed> $expected
-	 * @param array<array-key, mixed> $actual
-	 */
-	private function assertArraySubsetRecursive(array $expected, array $actual, string $message = '', string $path = '$'): void
-	{
-		foreach ($expected as $key => $value) {
-			$this->assertArrayHasKey($key, $actual, ltrim("{$message}\nMissing key at path: {$path}"));
-
-			if (!is_array($value) || !is_array($actual[$key])) {
-				$this->assertSame($value, $actual[$key], ltrim("{$message}\nMismatch at path: {$path}"));
-				continue;
-			}
-
-			$subPath = is_string($key) && preg_match('/^\w+$/', $key) === 1
-				? "{$path}.{$key}"
-				: "{$path}[" . var_export($key, true) . "]";
-			$this->assertArraySubsetRecursive($value, $actual[$key], $message, $subPath);
-		}
 	}
 }
