+++ tests/Unit/Fixer/BlockStringFixerTest.php
@@ -72,13 +72,13 @@ final class BlockStringFixerTest extends TestCase
 		yield 'nowdoc with unregistered delimiter should be left unchanged' => [
 			'config' => ['formatters' => []],
 			'input' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<'HTML'
 					<h1>Hello world!</h1>
 					HTML;
 				PHP,
 			'expected' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<'HTML'
 					<h1>Hello world!</h1>
 					HTML;
@@ -102,7 +102,7 @@ final class BlockStringFixerTest extends TestCase
 				],
 			],
 			'input' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<'HTML'
 					<h1>Hello world1</h1>
 					HTML;
@@ -114,7 +114,7 @@ final class BlockStringFixerTest extends TestCase
 					XML;
 				PHP,
 			'expected' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<'HTML'
 					Hello world1
 					HTML;
@@ -155,7 +155,7 @@ final class BlockStringFixerTest extends TestCase
 				],
 			],
 			'input' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<'HTML'
 					Hello world
 					HTML;
@@ -164,7 +164,7 @@ final class BlockStringFixerTest extends TestCase
 					XML;
 				PHP,
 			'expected' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<'HTML'
 					<htm>Hello world</htm>
 					HTML;
@@ -191,13 +191,13 @@ final class BlockStringFixerTest extends TestCase
 				],
 			],
 			'input' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<"HTML"
 					<h1 class="{$e['class']}">Hello $planet!</h1>
 					HTML;
 				PHP,
 			'expected' => <<<'PHP'
-				<?php
+				<?php declare(strict_types=1);
 				echo <<<"HTML"
 					Hello $planet!
 					HTML;
